pyproject.toml
[project]
name = "xfmr-zem"
version = "0.2.4"
description = "Zem: Unified Data Pipeline Framework (ZenML + NeMo Curator + DataJuicer) for multi-domain processing"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = "Apache-2.0"
authors = [
    {name = "Khai Hoang", email = "khaihq@vbiacademy.edu.vn"}
]
keywords = ["xfmr-zem", "data-pipeline", "zenml", "nemo-curator", "data-juicer", "mlops"]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "fastmcp>=0.1.0",
    "mcp>=0.1.0",
    "zenml[local,server]>=0.75.0",
    "pyarrow>=15.0.0",
    "nemo-curator>=1.0.0",
    "dask-cuda>=24.0.0",
    "ftfy>=6.3.1",
    "pydantic>=2.0.0",
    "datasketch>=1.6.0",
    "underthesea>=9.2.9",
    "sentence-transformers>=5.2.2",
    "faiss-cpu>=1.13.2",
    "transformers>=4.55.2",
    "unstructured[all-docs]>=0.16.0",
    "python-magic>=0.4.27",
]

[project.urls]
Homepage = "https://github.com/OAI-Labs/xfmr-zem"
Repository = "https://github.com/OAI-Labs/xfmr-zem"
Issues = "https://github.com/OAI-Labs/xfmr-zem/issues"
Changelog = "https://github.com/OAI-Labs/xfmr-zem/blob/main/CHANGELOG.md"

[project.optional-dependencies]
zenml = [
    "zenml>=0.75.0",
]
nemo = [
    "nemo-curator>=0.6.0",
]
datajuicer = [
    "py-data-juicer>=1.0.0",
]
deduplication = [
    "datasketch>=1.6.0",
    "underthesea>=6.8.0",
]
all = [
    "zenml>=0.75.0",
    "nemo-curator>=0.6.0",
    "py-data-juicer>=1.0.0",
    "datasketch>=1.6.0",
    "underthesea>=6.8.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
    "transformers>=4.40.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "pillow>=10.0.0",
    "onnxruntime>=1.16.0",
    "opencv-python>=4.8.0",
    "shapely",
    "pyclipper",
    "einops",
    "pdfplumber>=0.11.0",
    "ruamel.yaml>=0.17.0",
    "cachetools>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
zem = "xfmr_zem.cli:main"
xz = "xfmr_zem.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xfmr_zem"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]
